requirements.txt
opencv-python==4.6.0.66
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
easydict==1.9
scipy==1.9.3
numpy==1.23.4
matplotlib==3.6.1
tqdm==4.36.1
Pillow==9.2.0
tensorboardX==2.5.1
